QCN9274 hw2.0: 1.0.1: add WLAN.WBE.1.0.1-00029-QCAHKSWPL_SILICONZ-1
This upload request is for Ath12k 12.1.0 SP for Waikiki
IPQ9574.ATH.12.1.0 CS release for WKK V2 hardware
QCN9274 is PCIE based chipset which supports 2.4G, 5G and 6G wireless modes in 4x4 mode
with 320mhz bandwidth support.This firmware is for 11be based WKK hardware
and validated with basic 11ax mode in upstream driver. MLO is not validated with upstream driver yet.